==> 14_backtester_loop.ipynb <==
# Spikeet clone is for research and testing
# Backtester is for realistic testing (stoploss, ranking)

# INTRADAY BACKTESTER

# Loop through all days
    # Load the data for the day using the historical scanner
    # Loop through all minutes
        # Check for trading signal
            # Append trade log, research log, portfolio log, order log
    # Record portfolio value
